requirements.txt
pydantic
fastapi
python-dotenv
uvicorn
redis
motor
elasticsearch
sse_starlette